refactor: use mapper as a service
refactor: correctly split model from dto
