+++ exercices/steps/apps/nest-server/src/app/feature/cart/add-cart-item.dto.ts
@@ -11,4 +11,9 @@ export class AddCartItemDTO {
   @IsNotEmpty()
   @IsNumber()
   quantity: number;
+
+  @ApiProperty()
+  @IsNotEmpty()
+  @IsNumber()
+  price: number;
 }
