--- exercices/steps/apps/nest-server/src/app/feature/cart/cart-item.dto.ts
@@ -1,6 +1,0 @@
-import { Book } from "@nest-server/app/shared/models/book.model";
-
-export class CartItemDTO {
-    book: Book;
-    quantity: number;
-}
