+++ exercices/steps/apps/nest-server/src/app/feature/cart/cart.controller.ts
@@ -1,23 +1,22 @@
-import { Cart, CART_ITEM_SCHEMA } from '@nest-server/app/shared/models/basket.model';
+import { BookCartItem, CART_ITEM_SCHEMA } from '@nest-server/app/shared/models/cart.model';
 import { INTERNAL_SERVER_SCHEMA } from '@nest-server/app/shared/models/error.model';
 import { Body, Controller, Get, HttpCode, Put } from '@nestjs/common';
 import { ApiBody, ApiInternalServerErrorResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger';
-import { BooksService } from '../books/books.service';
 import { AddCartItemDTO } from './add-cart-item.dto';
-import { CartItemDTO } from './cart-item.dto';
+import { CartMapperService } from './cart-mapper.service';
 import { CartService } from './cart.service';
 
 @Controller('cart')
 @ApiTags('cart')
 export class BasketController {
-  constructor(private readonly cartService: CartService, private readonly booksService: BooksService) {}
+  constructor(private readonly cartService: CartService, private readonly cartMapperService: CartMapperService) {}
 
   @Get()
   @ApiInternalServerErrorResponse({ status: 500, description: 'Internal server error', schema: INTERNAL_SERVER_SCHEMA })
   @ApiOkResponse({ status: 200, description: 'The user basket', schema: { type: 'array', items: CART_ITEM_SCHEMA } })
-  async getBasket(): Promise<CartItemDTO[]> {
+  async getBasket(): Promise<BookCartItem[]> {
     const cart = await this.cartService.readBasket();
-    return this.mapCartToItemDTO(cart);
+    return this.cartMapperService.toCartItems(cart);
   }
 
   @Put()
@@ -27,16 +26,8 @@ export class BasketController {
   @ApiBody({ type: AddCartItemDTO })
   async addToBasket(
     @Body() cartItem: AddCartItemDTO
-  ): Promise<CartItemDTO[]> {
+  ): Promise<BookCartItem[]> {
     const cart = await this.cartService.writeItem(cartItem);
-    return this.mapCartToItemDTO(cart);
-  }
-
-  private mapCartToItemDTO(cart: Cart): CartItemDTO[] {
-    return this.booksService.getAll(Object.keys(cart))
-      .map(book => ({
-        book,
-        quantity: cart[book.id]
-      }));
+    return this.cartMapperService.toCartItems(cart);
   }
 }
