--- exercices/steps/apps/nest-server/src/app/feature/cart/cart.dto.ts
@@ -1,14 +1,0 @@
-import { CartItem } from '@nest-server/app/shared/models/basket.model';
-import { ApiProperty } from '@nestjs/swagger';
-import { IsNotEmpty, IsNumber } from 'class-validator';
-
-export class CartDTO {
-  @ApiProperty()
-  @IsNotEmpty()
-  books: CartItem[]
-
-  @ApiProperty()
-  @IsNotEmpty()
-  @IsNumber()
-  total: number;
-}
